Extract RestClient
