--- app/src/main/java/com/bogdan/codeforceswatcher/model/ContestResponse.kt
@@ -1,6 +1,0 @@
-package com.bogdan.codeforceswatcher.model
-
-data class ContestResponse(
-    val status: String,
-    val result: List<Contest>
-)
