--- app/src/main/java/com/bogdan/codeforceswatcher/model/RatingChangeResponse.kt
@@ -1,6 +1,0 @@
-package com.bogdan.codeforceswatcher.model
-
-data class RatingChangeResponse(
-    val status: String,
-    val result: List<RatingChange>
-)
