--- app/src/main/java/com/bogdan/codeforceswatcher/model/UserResponse.kt
@@ -1,6 +1,0 @@
-package com.bogdan.codeforceswatcher.model
-
-data class UserResponse(
-    val status: String,
-    val result: List<User>
-)
