+++ app/src/main/java/com/bogdan/codeforceswatcher/network/model/ContestResponse.kt
@@ -1,0 +1,8 @@
+package com.bogdan.codeforceswatcher.network.model
+
+import com.bogdan.codeforceswatcher.model.Contest
+
+data class ContestResponse(
+    val status: String,
+    val result: List<Contest>
+)
