+++ app/src/main/java/com/bogdan/codeforceswatcher/network/model/RatingChangeResponse.kt
@@ -1,0 +1,8 @@
+package com.bogdan.codeforceswatcher.network.model
+
+import com.bogdan.codeforceswatcher.model.RatingChange
+
+data class RatingChangeResponse(
+    val status: String,
+    val result: List<RatingChange>
+)
