+++ app/src/main/java/com/bogdan/codeforceswatcher/network/model/UserResponse.kt
@@ -1,0 +1,8 @@
+package com.bogdan.codeforceswatcher.network.model
+
+import com.bogdan.codeforceswatcher.model.User
+
+data class UserResponse(
+    val status: String,
+    val result: List<User>
+)
